requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
websockets==12.0
python-binance==1.0.19
pydantic-settings==2.4.0